+++ allauth_2fa/adapter.py
@@ -3,8 +3,11 @@ try:
 except ImportError:
     from urllib import urlencode
 
-from django.core.urlresolvers import reverse
 from django.http import HttpResponseRedirect
+try:
+    from django.urls import reverse
+except ImportError:
+    from django.core.urlresolvers import reverse
 
 from allauth.exceptions import ImmediateHttpResponse
 from allauth.account.adapter import DefaultAccountAdapter