+++ allauth_2fa/middleware.py
@@ -1,13 +1,17 @@
 from django.shortcuts import redirect
 from django.conf import settings
 from django.contrib import messages
-from django.core.urlresolvers import resolve, reverse
+try:
+    from django.urls import resolve, reverse
+except ImportError:
+    from django.core.urlresolvers import resolve, reverse
 try:
     from django.utils.deprecation import MiddlewareMixin
 except ImportError:
     MiddlewareMixin = object
 
 from allauth.account.adapter import get_adapter
+from allauth.compat import is_anonymous
 
 
 class AllauthTwoFactorMiddleware(MiddlewareMixin):
@@ -90,7 +94,7 @@ class BaseRequire2FAMiddleware(MiddlewareMixin):
 
     def process_view(self, request, view_func, view_args, view_kwargs):
         # The user is not logged in, do nothing.
-        if request.user.is_anonymous():
+        if is_anonymous(request.user):
             return
 
         # If this doesn't require 2FA, then stop processing.