+++ CHANGELOG.rst
@@ -12,8 +12,9 @@ Changelog
   without enabling 2FA first.
 * Add base middleware to ensure particular users (e.g. superusers) have 2FA
   enabled.
-* Drop support for Django 1.9 and 1.10, they're
-  [no longer supported](https://www.djangoproject.com/download/#supported-versions).
+* Drop official support for Django 1.9 and 1.10, they're
+  [no longer supported](https://www.djangoproject.com/download/#supported-versions)
+  by the Django project.
 
 0.4.4 March 24, 2017
 ====================